# SQLite.demo-master/api-docs.yaml
openapi: 3.0.1
info:
  title: Product iiiii
  description: product services
  version: "2.0"
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /deletestudent:
    delete:
      tags:
      - student-controller
      operationId: deleteStudent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
...
  /createstudent:
    post:
      tags:
      - student-controller
      operationId: createStudent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
...
components:
  schemas:
    Student:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        email:
          type: string
...